requirements.txt
functions-framework==3.*
google-cloud-storage
pdf2docx
datetime
google-auth